requirements.txt
pip install tkinter
pip install folium
pip install pywebview
pip install customtkinter
pip install fastapi uvicorn websockets
pip install folium pywebview tkinter customtkinter
fastapi==0.68.1
uvicorn==0.15.0
websockets==10.0
pyttsx3==2.90
SpeechRecognition==3.8.1
pyserial==3.5
pynmea2==1.18.0
folium==0.12.1
pywebview==3.6.3
customtkinter==4.6.3
psutil==5.9.0
PyYAML==6.0
requests==2.26.0
geopy==2.2.0